requirements.txt
aioredis==1.3.1
alabaster==0.7.12
amqp==2.6.0
aniso8601==7.0.0
argon2-cffi==20.1.0
arrow==0.12.1
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
Babel==2.8.0
bcrypt==3.1.7
beautifulsoup4==4.6.3
behave==1.2.6
behave-django==1.1.0
billiard==3.5.0.5
bleach==3.3.0
blessings==1.7
blinker==1.4
boto==2.49.0
boto3==1.14.52
botocore==1.17.52
bpython==0.17.1
cachetools==4.1.1
celery==4.2.1
certifi==2020.6.20
cffi==1.14.2
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
constantly==15.1.0
coverage==4.5.1
daphne==2.5.0
dj-database-url==0.5.0
Django==2.2.13
django-allauth==0.42.0
django-anymail==7.2.1
django-appconf==1.0.4
django-autotranslate==1.1.1
django-behave==0.1.6
django-blog-it==0.3
django-bootstrap-form==3.4
django-celery-beat==1.2.0
django-compressor==2.2
django-contrib-comments==1.9.2
django-cors-headers==3.3.0
django-countries==7.1
django-crispy-forms==1.9.2
django-crontab==0.7.1
django-debug-toolbar==1.10.1
django-debug-toolbar-template-timings==0.9
django-endless-pagination==2.0
django-environ==0.4.5
django-extra-views==0.13.0
django-haystack==3.0
django-helpdesk==0.2.22
django-import-export==2.5.0
django-maintenancemode-2==1.2
django-markdown-deux==1.0.5
django-markdownx==3.0.1
django-oscar==3.0.2
django-phonenumber-field==3.0.1
django-redis
django-rest-auth==0.9.5
django-simple-pagination==1.3
django-storages==1.9.1
django-tables2==2.3.4
django-taggit==1.3.0
django-tellme==0.6.5
django-timezone-field==4.0
django-treebeard==4.4
django-web-profiler==0.1.4
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
dodgy==0.2.1
elasticsearch==2.3.0
email-reply-parser==0.5.9
entrypoints==0.3
et-xmlfile==1.0.1
factory-boy==2.12.0
Faker==8.0.0
fasteners==0.14.1
ftfy==5.9
future==0.18.2
futures==3.1.1
gevent==20.6.2
google-api-python-client==1.7.4
google-auth==1.20.0
google-auth-httplib2==0.0.4
goslate==1.5.1
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==0.4.16
gunicorn==20.1.0
hiredis==1.1.0
httplib2==0.18.1
hyperlink==20.0.1
idna==2.10
imagesize==1.2.0
incremental==17.5.0
isort==4.3.21
Jinja2==2.11.2
jmespath==0.10.0
jsonpickle==1.0
keyring==18.0.1
kombu==4.6.11
lxml==4.5.2
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
markdown2==2.4.0
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
microurl==0.1.1
monotonic==1.5
msgpack==1.0.0
netifaces==0.10.4
numpy
oauthlib==3.1.0
opencv-python
phonenumbers==8.12.20
Pillow==7.2.0
prospector==1.1.4
protobuf==3.6.1
psycopg2-binary
purl==1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
pylint-plugin-utils==0.6
pymacaroons==0.13.0
pymemcache==2.0.0
pymongo==3.7.2
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyRFC3339==1.1
pytesseract==0.3.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-magic==0.4.15
python-memcached==1.59
python-slugify==4.0.1
python3-openid==3.2.0
raven==6.9.0
rcssmin==1.0.6
redis==2.10.6
reportlab==3.5.34
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
requirements-detector==0.7
rjsmin==1.0.12
rsa==4.6
Rx==1.6.1
s3transfer==0.3.3
SecretStorage==2.3.1
selenium==3.141.0
sendgrid==5.6.0
service-identity==18.1.0
setoptconf==0.2.0
simplejson==3.17.2
singledispatch==3.4.0.3
six==1.15.0
snowballstemmer==2.0.0
sorl-thumbnail==12.5.0
Sphinx==3.1.2
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
tablib==3.0.0
text-unidecode==1.3
tinys3==0.1.12
toml==0.10.1
Twisted==20.3.0
twython==3.7.0
txaio==20.4.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.10
utils==0.9.0
webencodings==0.5.1
whitenoise==5.2.0
zope.event==4.4
zope.interface==5.1.0